refactor: OrderTables 일급 객체 추가
